test: [NUTR-107] - Updated new utils
